+++ lib/service/contact-service.ts
@@ -5,9 +5,10 @@ import ContactDao from "../dao/contact-dao";
 import SigningKeysDao from "../dao/signing-keys-dao";
 import { Client, Contact, SigningKey, Tenant } from "@/graphql/generated/graphql-types";
 import { randomUUID } from "crypto";
-import { CONTACT_TYPE_FOR_CLIENT, CONTACT_TYPE_FOR_SIGNING_KEY, CONTACT_TYPE_FOR_TENANT } from "@/utils/consts";
+import { CLIENT_READ_SCOPE, CLIENT_UPDATE_SCOPE, CONTACT_TYPE_FOR_CLIENT, CONTACT_TYPE_FOR_SIGNING_KEY, CONTACT_TYPE_FOR_TENANT, KEY_READ_SCOPE, KEY_UPDATE_SCOPE, TENANT_READ_ALL_SCOPE, TENANT_READ_SCOPE, TENANT_UPDATE_SCOPE } from "@/utils/consts";
 import { GraphQLError } from "graphql";
 import { DaoFactory } from "../data-sources/dao-factory";
+import { authorizeByScopeAndTenant, ServiceAuthorizationWrapper } from "@/utils/authz-utils";
 
 
 const tenantDao: TenantDao = DaoFactory.getInstance().getTenantDao();
@@ -25,28 +26,82 @@ class ContactService {
     }
 
     public async getContacts(objectId: string): Promise<Array<Contact>> {
-        return contactDao.getContacts(objectId);
+        const getData = ServiceAuthorizationWrapper(
+            {
+                async performOperation(_, __) {
+                    return contactDao.getContacts(objectId);
+                },
+                async additionalConstraintCheck(oidcContext: OIDCContext, result: Array<Contact> | null): Promise<{ isAuthorized: boolean, errorMessage: string | null, result: Array<Contact> | null }> {
+                    if (result && result.length > 0) {
+                        // just need the first contact in the list, since they will all by
+                        // tied to the same type of object, based on the object id.
+                        const contact: Contact = result[0];
+                        let scopeRequired: string = "";
+                        let tenantId: string = "";
+                        if (contact.objecttype === CONTACT_TYPE_FOR_TENANT) {
+                            const tenant: Tenant | null = await tenantDao.getTenantById(contact.objectid);
+                            if (!tenant) {
+                                throw new GraphQLError("ERROR_TENANT_DOES_NOT_EXIST");
+                            }
+                            tenantId = tenant.tenantId;
+                            scopeRequired = TENANT_READ_SCOPE;
+                        }
+                        else if (contact.objecttype === CONTACT_TYPE_FOR_CLIENT) {
+                            const client: Client | null = await clientDao.getClientById(contact.objectid);
+                            if (!client) {
+                                throw new GraphQLError("ERROR_CLIENT_DOES_NOT_EXIST");
+                            }
+                            tenantId = client.tenantId;
+                            scopeRequired = CLIENT_READ_SCOPE;
+                        }
+                        else if (contact.objecttype === CONTACT_TYPE_FOR_SIGNING_KEY) {
+                            const key: SigningKey | null = await keyDao.getSigningKeyById(contact.objectid);
+                            if (!key) {
+                                throw new GraphQLError("ERROR_KEY_DOES_NOT_EXIST");
+                            }
+                            tenantId = key.tenantId;
+                            scopeRequired = KEY_READ_SCOPE;
+                        }
+                        const {isAuthorized, errorMessage} = authorizeByScopeAndTenant(oidcContext, [TENANT_READ_ALL_SCOPE, scopeRequired], tenantId);
+                        if(!isAuthorized){
+                            return {isAuthorized, errorMessage, result: null};
+                        }
+                        return {isAuthorized: true, errorMessage: null, result: result};
+                    }
+                    return {isAuthorized: true, errorMessage: null, result: result};
+                }
+            }
+        );
+        const contacts = await getData(this.oidcContext, [TENANT_READ_ALL_SCOPE, CLIENT_READ_SCOPE, TENANT_READ_SCOPE, KEY_READ_SCOPE], objectId);
+        return contacts || [];
     }
 
     public async addContact(contact: Contact): Promise<Contact> {
-
+        let scopeRequired: string = "";
+        let tenantId: string = "";
         if(contact.objecttype === CONTACT_TYPE_FOR_TENANT){
             const tenant: Tenant | null = await tenantDao.getTenantById(contact.objectid);
             if(!tenant){
                 throw new GraphQLError("ERROR_TENANT_DOES_NOT_EXIST");
             }
+            tenantId = tenant.tenantId;
+            scopeRequired = TENANT_UPDATE_SCOPE;
         }
         else if(contact.objecttype === CONTACT_TYPE_FOR_CLIENT){
             const client: Client | null = await clientDao.getClientById(contact.objectid);
             if(!client){
                 throw new GraphQLError("ERROR_CLIENT_DOES_NOT_EXIST");
             }
+            tenantId = client.tenantId;
+            scopeRequired = CLIENT_UPDATE_SCOPE;
         }
         else if(contact.objecttype === CONTACT_TYPE_FOR_SIGNING_KEY){
             const key: SigningKey | null = await keyDao.getSigningKeyById(contact.objectid);
             if(!key){
                 throw new GraphQLError("ERROR_KEY_DOES_NOT_EXIST");
             }
+            tenantId = key.tenantId;
+            scopeRequired = KEY_UPDATE_SCOPE;
         }
         else {
             throw new GraphQLError("ERROR_INVALID_CONTACT_TYPE")
@@ -55,13 +110,48 @@ class ContactService {
             throw new GraphQLError("ERROR_INVALID_EMAIL");
         }
 
+        const {isAuthorized, errorMessage} = authorizeByScopeAndTenant(this.oidcContext, scopeRequired, tenantId);
+        if(!isAuthorized){
+            throw new GraphQLError(errorMessage || "ERROR");
+        }
+
         contact.contactid = randomUUID().toString();
         await contactDao.addContact(contact);
         return Promise.resolve(contact);        
     }
 
     public async removeContact(contactId: string): Promise<void> {
-        await contactDao.removeContact(contactId);
+        const contact: Contact | null = await contactDao.getContactById(contactId);
+        if(contact){
+            let scopeRequired: string = "";
+            let tenantId: string = "";
+            if(contact.objecttype === CONTACT_TYPE_FOR_TENANT){
+                tenantId = contact.objectid;
+                scopeRequired = TENANT_UPDATE_SCOPE;
+            }
+            else if(contact.objecttype === CONTACT_TYPE_FOR_CLIENT){
+                const client: Client | null = await clientDao.getClientById(contact.objectid);
+                if(!client){
+                    throw new GraphQLError("ERROR_CLIENT_DOES_NOT_EXIST");
+                }
+                tenantId = client.tenantId;
+                scopeRequired = CLIENT_UPDATE_SCOPE;
+            }
+            else if(contact.objecttype === CONTACT_TYPE_FOR_SIGNING_KEY){
+                const key: SigningKey | null = await keyDao.getSigningKeyById(contact.objectid);
+                if(!key){
+                    throw new GraphQLError("ERROR_KEY_DOES_NOT_EXIST");
+                }
+                tenantId = key.tenantId;
+                scopeRequired = KEY_UPDATE_SCOPE;
+            }
+            const {isAuthorized, errorMessage} = authorizeByScopeAndTenant(this.oidcContext, scopeRequired, tenantId);
+            if(!isAuthorized){
+                throw new GraphQLError(errorMessage || "ERROR");
+            }
+            await contactDao.removeContact(contactId);
+        }
+        return Promise.resolve();        
     }
 }
